% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GPDxgb.cv}
\alias{GPDxgb.cv}
\title{Cross-validation for a Generalized Pareto (GPD) boosting model with shape parameter xi>=0. Grid search for xi.}
\usage{
GPDxgb.cv(
  y,
  X,
  xi,
  simultaneous = FALSE,
  init.sigma,
  stepsize = 0.1,
  stepsize_xi = 0.1,
  tree_depth = 5,
  tree_depth_xi = 3,
  nrounds,
  xis = NULL,
  cv.nfold = 5,
  nthread = NULL,
  orthogonal = FALSE
)
}
\arguments{
\item{y}{a vector of observations of length \code{n}, or a matrix with \code{n} rows.}

\item{X}{a matrix or data.frame object. Design matrix of dimension \code{n * p}.}

\item{xi}{a non-negative numeric. The shape parameter of the GPD. Also indicates the starting value of xi in the boosting algorithm if \code{simultaneous=TRUE}. Note, if \code{xi=0}, the distribution is of type exponential}

\item{simultaneous}{a logical. Should simultaneous boosting of the the sigma and xi parameters be performed? If \code{FALSE}, the xi parameter will be fixed. By default, \code{FALSE}.}

\item{init.sigma}{a positive numeric. Indicates the starting value of sigma in the boosting algorithm.}

\item{stepsize}{a positive numeric. \code{learning_rate} parameter for the boosting model for sigma.
See \link[xgboost]{xgb.train} for more details.}

\item{stepsize_xi}{a positive numeric. \code{learning_rate} parameter for the boosting model for xi.
See \link[xgboost]{xgb.train} for more details.}

\item{tree_depth}{a positive numeric. \code{tree_depth} parameter for the boosting model for sigma.
See \link[xgboost]{xgb.train} for more details.}

\item{tree_depth_xi}{a positive numeric. \code{tree_depth} parameter for the boosting model for xi.
See \link[xgboost]{xgb.train} for more details.}

\item{nrounds}{a positive scalar \code{nrounds} parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{xis}{a non-negative vector. Will be ignored if \code{simultaneous=TRUE}. If not supplied when \code{simultaneous=FALSE}, the argument \code{xi} will be used instead.}

\item{cv.nfold}{a positive scalar. Determines the number of folds for cross-validation.}

\item{nthread}{a positive scalar. Number of thread used for cross-validation. If not supplied, all available cores are used}

\item{orthogonal}{a logical. Should the two parameters be orthogonalized. By default, \code{FALSE}.}
}
\value{
a list
}
\description{
Cross-validation for a Generalized Pareto (GPD) boosting model with shape parameter xi>=0. Grid search for xi.
}
