% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GPDxgb.train}
\alias{GPDxgb.train}
\title{Title}
\usage{
GPDxgb.train(
  y,
  X,
  xi = 0.1,
  simultaneous = FALSE,
  init.sigma,
  stepsize = 0.1,
  tree_depth = 5,
  tree_depth_xi = 3,
  rounds
)
}
\arguments{
\item{rounds}{}
}
\description{
Title
}
