% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GPDxgb.train}
\alias{GPDxgb.train}
\title{Fit a Generalized Pareto (GPD) boosting model with shape parameter xi>=0}
\usage{
GPDxgb.train(
  y,
  X,
  xi,
  simultaneous = FALSE,
  init.sigma,
  stepsize = 0.1,
  stepsize_xi = 0.1,
  tree_depth = 5,
  tree_depth_xi = 3,
  nrounds,
  X_test = NULL,
  orthogonal = FALSE
)
}
\arguments{
\item{y}{a vector of observations of length \code{n}, or a matrix with \code{n} rows.}

\item{X}{a matrix or data.frame object. Design matrix of dimension \code{n * p}.}

\item{xi}{a non-negative numeric. The shape parameter of the GPD. Also indicates the starting value of xi in the boosting algorithm if \code{simultaneous=TRUE}. Note, if \code{xi=0}, the distribution is of type exponential}

\item{simultaneous}{a logical. Should simultaneous boosting of the the sigma and xi parameters be performed? If \code{FALSE}, the xi parameter will be fixed. By default, \code{FALSE}.}

\item{init.sigma}{a positive numeric. Indicates the starting value of sigma in the boosting algorithm.}

\item{stepsize}{a positive scalar \code{learning_rate} parameter for the boosting model for sigma.
See \link[xgboost]{xgb.train} for more details.}

\item{stepsize_xi}{a positive numeric. \code{learning_rate} parameter for the boosting model for xi.
See \link[xgboost]{xgb.train} for more details.}

\item{tree_depth}{a positive numeric. \code{tree_depth} parameter for the boosting model for sigma.
See \link[xgboost]{xgb.train} for more details.}

\item{tree_depth_xi}{a positive numeric. \code{tree_depth} parameter for the boosting model for xi.
See \link[xgboost]{xgb.train} for more details.}

\item{nrounds}{a positive numeric. \code{nrounds} parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{X_test}{a matrix or data.frame object. Design matrix with the same row names as \code{X} to perform predictions on. By default set to \code{X}.}

\item{orthogonal}{a logical. Should the two parameters be orthogonalized. By default, \code{FALSE}.}
}
\value{
a list
}
\description{
Fit a Generalized Pareto (GPD) boosting model with shape parameter xi>=0
}
